+++ src/vue/pages/democracy/tabs/dispatch/ScheduledList.vue
@@ -1,11 +1,38 @@
 <template>
   <div class="scheduled-list">
-    <h1 class="scheduled-list__header">
-      {{ $t('democracy-page.scheduled-list.scheduled-header') }}
-    </h1>
-    <template v-if="true">
-      <template v-if="false">
-        <div class="scheduled-list__body"/>
+    <div
+      class="scheduled-list__headers"
+      :class="{'scheduled-list__headers--grid': entriesFiltered?.length}"
+    >
+      <h1>
+        {{ $t('democracy-page.scheduled-list.scheduled-header') }}
+      </h1>
+      <template v-if="entriesFiltered?.length">
+        <h4>
+          {{ $t('democracy-page.scheduled-list.id-header') }}
+        </h4>
+        <h4>
+          {{ $t('democracy-page.scheduled-list.remaining-header') }}
+        </h4>
+        <h4>
+          {{ $t('democracy-page.scheduled-list.period-header') }}
+        </h4>
+        <h4>
+          {{ $t('democracy-page.scheduled-list.count-header') }}
+        </h4>
+      </template>
+    </div>
+    <template v-if="entriesFiltered">
+      <template v-if="entriesFiltered.length">
+        <template
+          v-for="(entry, index) in entriesFiltered"
+          :key="index"
+        >
+          <schedule-row
+            class="scheduled-list__row"
+            :entry="entry"
+          />
+        </template>
       </template>
       <template v-else>
         <no-data-message
@@ -22,15 +49,72 @@
 </template>
 
 <script>
+import ScheduleRow from '@democracy-page/tabs/dispatch/ScheduleRow'
 import SkeletonLoader from '@/vue/common/SkeletonLoader'
 import NoDataMessage from '@/vue/common/NoDataMessage'
 
+import { computed } from 'vue'
+import { useCall } from '@/vue/composables'
+import { api } from '@api'
 export default {
   name: 'scheduled-list',
 
   components: {
+    ScheduleRow,
     SkeletonLoader,
     NoDataMessage,
+  },
+
+  setup () {
+    const transformEntries = {
+      transform: (entries) => {
+        return entries
+          .filter(([_, vecSchedOpt]) =>
+            vecSchedOpt.some((schedOpt) => schedOpt.isSome))
+          .reduce((items, [key, vecSchedOpt]) => {
+            const blockNumber = key.args[0]
+
+            return vecSchedOpt.reduce((acc, schedOpt) => {
+              if (schedOpt.isSome) {
+                const {
+                  call,
+                  maybeId,
+                  maybePeriodic,
+                  priority,
+                } = schedOpt.unwrap()
+
+                acc.push({
+                  blockNumber,
+                  call,
+                  maybeId,
+                  maybePeriodic,
+                  priority,
+                })
+              }
+
+              return acc
+            }, items)
+          }, [])
+      },
+    }
+
+    const bestNumber = useCall(api.derive.chain.bestNumber)
+
+    const entries = useCall(
+      api.query.scheduler.agenda.entries,
+      undefined,
+      transformEntries,
+    )
+
+    const entriesFiltered = computed(() => {
+      if (!bestNumber.value || !entries.value) return null
+      return entries.value
+        .filter(({ blockNumber }) => blockNumber.gte(bestNumber.value))
+    })
+
+    return {
+      entriesFiltered,
+    }
   },
 }
 </script>
@@ -39,9 +123,23 @@ export default {
 @import '~@scss/mixins';
 @import '~@scss/variables';
 
-.scheduled-list__header {
-  padding: 0 1.6rem;
-  margin-bottom: 2rem;
+.scheduled-list {
+  overflow-x: auto;
+
+  @include scrollbar;
 }
 
+.scheduled-list__headers {
+  margin-bottom: 2rem;
+
+  &--grid {
+    @include democracy-schedule-grid-row(flex-end);
+  }
+}
+
+.scheduled-list__row {
+  & + & {
+    margin-top: 0.4rem;
+  }
+}
 </style>